`@repo/eslint-config` source files moved to `src` directory
