+++ packages/eslint-config/package.json
@@ -4,9 +4,9 @@
   "version": "0.0.0",
   "type": "module",
   "exports": {
-    "./base": "./base.ts",
-    "./next": "./next.ts",
-    "./react": "./react.ts"
+    "./base": "./src/base.ts",
+    "./next": "./src/next.ts",
+    "./react": "./src/react.ts"
   },
   "scripts": {
     "check-types": "tsc --noEmit",
