+++ packages/eslint-config/src/next.ts
