+++ packages/eslint-config/src/react.ts
